refactor: videoCard 요소 불필요한 클래스 선언 제거
단순 템플릿이라고 생각했습니다.
